move pluginregistry to its own dir
