+++ pkg/framework/pluginregistry/pluginregistry.go
